SplitGenerator: minor api change

+++ geotrellis-spark/src/main/scala/geotrellis/spark/rdd/SplitGenerator.scala
@@ -21,7 +21,7 @@ case class RasterSplitGenerator(
 }
 
 object RasterSplitGenerator {
-  def apply(tileBounds: TileBounds, zoom: Int, tileSizeBytes: Int, blockSizeBytes: Int) = {
+  def apply(tileBounds: TileBounds, zoom: Int, tileSizeBytes: Int, blockSizeBytes: Long) = {
     new RasterSplitGenerator(tileBounds, zoom, computeIncrement(tileBounds, tileSizeBytes, blockSizeBytes))
   }
   def computeIncrement(tileBounds: TileBounds, tileSizeBytes: Int, blockSizeBytes: Long) = {
